requirements.txt
torch==1.12.1
numpy==1.23.4
Pillow==9.2.0
basicsr==1.4.2
-e git+https://github.com/CompVis/stable-diffusion.git@69ae4b35e0a0f6ee1af8bb9a5d0016ccb27e36dc#egg=latent-diffusion
aiohttp==3.8.3
pydantic==1.10.2
lark==1.1.3
omegaconf==2.2.3
einops==0.5.0
transformers==4.23.1
pytorch_lightning==1.7.7
clip @ git+https://github.com/openai/CLIP@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
kornia==0.6.8
